make class shape abstract and extract render method
